XRT-604: updated analog ouput present value to be float as it should be

+++ src/bacnet/basic/object/ao.h
@@ -38,7 +38,7 @@ extern "C" {
 #endif /* __cplusplus */
 
     typedef struct analog_output_descr {
-        uint8_t Level[BACNET_MAX_PRIORITY];
+        float Level[BACNET_MAX_PRIORITY];
         bool Out_Of_Service;
         char *Name;
     } ANALOG_OUTPUT_DESCR;
